+++ src/Sciensoft.Hateoas/Filters/HateoasResultFilterAttribute.cs
@@ -3,7 +3,6 @@ using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.Extensions.Logging;
 using Sciensoft.Hateoas.Providers;
 using System;
-using System.Text.Json;
 using System.Threading.Tasks;
 
 namespace Sciensoft.Hateoas.Filters
@@ -12,23 +11,20 @@ namespace Sciensoft.Hateoas.Filters
 	{
 		private readonly ILogger<HateoasResultFilterAttribute> _logger;
 		private readonly IHateoasResultProvider _resultProvider;
-		private readonly JsonSerializerOptions _jsonOptions;
 
 		public HateoasResultFilterAttribute(
 			ILogger<HateoasResultFilterAttribute> logger,
-			IHateoasResultProvider resultProvider,
-			JsonSerializerOptions jsonOptions = null)
+			IHateoasResultProvider resultProvider)
 		{
 			_logger = logger ?? throw new ArgumentNullException(nameof(logger));
 			_resultProvider = resultProvider ?? throw new ArgumentNullException(nameof(resultProvider));
-			_jsonOptions = jsonOptions;
 		}
 
 		public override async Task OnResultExecutionAsync(ResultExecutingContext context, ResultExecutionDelegate next)
 		{
 			try
 			{
-				if (context.Result is ObjectResult result)
+				if (_resultProvider.HasAnyPolicy(context.Result, out ObjectResult result))
 				{
 					var finalResult = await _resultProvider.GetContentResultAsync(result).ConfigureAwait(false);
 					if (finalResult != null)