+++ src/Sciensoft.Hateoas/Providers/IHateoasResultProvider.cs
@@ -5,6 +5,8 @@ namespace Sciensoft.Hateoas.Providers
 {
 	internal interface IHateoasResultProvider
 	{
+		bool HasAnyPolicy(IActionResult actionResult, out ObjectResult objectResult);
+
 		Task<IActionResult> GetContentResultAsync(ObjectResult result);
 	}
 }